+++ packages/react-mui-audit-log-viewer/CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.28] - 2023-03-22
+
+### Fixed
+
+- Add errors support to audit log viewer
+
+## [0.0.27] - 2023-03-21
+
+### Fixed
+
+- Update to support config parameter to fetch custom audit log schema
+
+## [0.0.26] - 2023-03-20
+
+### Fixed
+
+- Update audit schema to snakecase ui_default_visible
+
 ## [0.0.25] - 2023-03-20
 
 ### Fixed